Personal Study Syllabus
Created to offer me a formal learning path
similar to that found in a degree course for maths and computer science.
At time of writing, I've just begun drafting it. It will be intended to
cover use of F#, Matlab and R rather than Java. Other languages will be
used to round out the skill sets.

After examining one or two syllabuses at local universities I came to the
conclusion that I have a fair idea of what I don't know and need a route
of my own to fill in knowledge that I either need for my own projects or
may require for more interesting career oppurtunities.

MOOC's and online sources will be heavily relied upon and referenced in
the syllabus for any interested readers in time to come.

The content is totally my opinion and not a qualified recommendation to
other users. Follow it at your own perils!

Additions, alternatives and suggestions very much welcome.